Adjust script to generate periodic tables
A few small changes are made to make the functions in `generate_periodic_tables` easier usable in jupyter notebooks.

Moreover, instead of only using global variables, the ones that might be changed by the user are now generated by a function `_get_config` and stored as a dictionary. The plotting functions expect this dictionary as an additional input now. In this way, it's easier to modify certain parameters when importing the functions from this script. This is especially useful for the widget that was created to simplify the generation of the periodic tables for users: https://github.com/t-reents/eos-plot-widget